+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added some external links related to Rolens to the application menu.
 * Fix an infinite loop bug when a document in the find view has been double clicked.
+* Find view: added the ability to make changes to single documents from within the object editor.
 
 ## [v0.1.0](https://github.com/garraflavatra/rolens/releases/tag/v0.1.0)